aileadgen-api/pyproject.toml
[project]
name = "aileadgen-api"
version = "0.1.0"
description = "AI Lead Gen FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib>=1.7.4",
  "python-multipart>=0.0.9",
  "email-validator>=2.2.0",
  "psycopg2-binary>=2.9.9",
]

[tool.setuptools]
packages = ["app"]
